shared MockResult
